Sunulation Builder refactor
