+++ simulations/SIMULATION-inf-dis-1.0-states-uninformed-0.2informed-0.2informing-0.2disinformed-0.2disinforming-0.2_.json
@@ -1,0 +1,1 @@
+{"name":"SIMULATION-inf-dis-1.0-states-uninformed-0.2informed-0.2informing-0.2disinformed-0.2disinforming-0.2_","nodeRoles":[{"coverage":1.0,"minCoverage":1,"nodeRole":{"name":"inf-dis","description":"inf-dis","transitionMap":[{"transitionType":"conditionProbability","sourceState":{"name":"uninformed"},"destinationState":{"name":"informed"},"probability":0.2,"provocativeNeighborName":["informing"]},{"transitionType":"conditionProbability","sourceState":{"name":"uninformed"},"destinationState":{"name":"disinformed"},"probability":0.25,"provocativeNeighborName":["disinforming"]},{"transitionType":"conditionProbability","sourceState":{"name":"informed"},"destinationState":{"name":"informing"},"probability":0.1,"provocativeNeighborName":["uninformed","disinformed"]},{"transitionType":"noConditionProbability","sourceState":{"name":"informing"},"destinationState":{"name":"informed"},"probability":1.0},{"transitionType":"conditionProbability","sourceState":{"name":"disinformed"},"destinationState":{"name":"disinforming"},"probability":0.1,"provocativeNeighborName":["uninformed","informed"]},{"transitionType":"noConditionProbability","sourceState":{"name":"disinforming"},"destinationState":{"name":"disinformed"},"probability":1.0},{"transitionType":"conditionProbability","sourceState":{"name":"disinformed"},"destinationState":{"name":"informed"},"probability":0.05,"provocativeNeighborName":["informing"]},{"transitionType":"conditionProbability","sourceState":{"name":"informed"},"destinationState":{"name":"disinformed"},"probability":0.05,"provocativeNeighborName":["disinforming"]}]},"nodeStates":[{"coverage":0.2,"minCoverage":1,"nodeState":{"name":"uninformed"},"color":{"argb":"ff000000"}},{"coverage":0.2,"minCoverage":1,"nodeState":{"name":"informed"},"color":{"argb":"ff000000"}},{"coverage":0.2,"minCoverage":1,"nodeState":{"name":"informing"},"color":{"argb":"ff000000"}},{"coverage":0.2,"minCoverage":1,"nodeState":{"name":"disinformed"},"color":{"argb":"ff000000"}},{"coverage":0.2,"minCoverage":1,"nodeState":{"name":"disinforming"},"color":{"argb":"ff000000"}}]}]}
